[ refactor ] Generalize type of prettyTCM
