+++ Agda.cabal
@@ -449,6 +449,8 @@ library
                     Agda.TypeChecking.Positivity
                     Agda.TypeChecking.Positivity.Occurrence
                     Agda.TypeChecking.Pretty
+                    Agda.TypeChecking.Pretty.Call
+                    Agda.TypeChecking.Pretty.Warning
                     Agda.TypeChecking.Primitive
                     Agda.TypeChecking.Primitive.Cubical
                     Agda.TypeChecking.ProjectionLike
