+++ src/full/Agda/TypeChecking/Pretty/Warning.hs-boot
@@ -1,0 +1,7 @@
+module Agda.TypeChecking.Pretty.Warning where
+
+import Agda.TypeChecking.Monad.Base
+import {-# SOURCE #-} Agda.TypeChecking.Pretty (PrettyTCM)
+
+instance PrettyTCM Warning
+
